Clean up imports after changes.
