+++ nimare/dataset/__init__.py
@@ -1,6 +1,7 @@
 """
 Top-level namespace for datasets.
 """
-from ..base import Contrast
 from .dataset import Database, Dataset
-from .extract import convert_sleuth, DataSource
+from .extract import DataSource
+
+__all__ = ['Database', 'Dataset', 'DataSource']
