+++ nimare/meta/cbma/kernel.py
@@ -10,10 +10,10 @@ from __future__ import division
 import numpy as np
 import nibabel as nib
 
-from nimare.utils import vox2mm
 from nilearn.image import resample_to_img, math_img
-from ...base import KernelTransformer
 from .utils import compute_ma, get_ale_kernel, peaks2maps
+from ...base.meta import KernelTransformer
+from ...utils.utils import vox2mm
 
 __all__ = ['ALEKernel', 'MKDAKernel', 'KDAKernel', 'Peaks2MapsKernel']
