+++ nimare/utils/__init__.py
@@ -1,11 +1,11 @@
 """
 Utility functions related to input/output, statistics, and general utilities.
 """
-from .io import convert_sleuth
+#from .io import convert_sleuth
 from .stats import one_way
 from .utils import get_template
 
 
 __all__ = [
-    'convert_sleuth', 'one_way', 'get_template',
+    'one_way', 'get_template',
 ]
