+++ nimare/utils/io.py
@@ -13,8 +13,8 @@ import pandas as pd
 import numpy as np
 from pyneurovault import api
 
-from nimare.dataset import Database
-from ..utils import get_resource_path, tal2mni
+from ..dataset import Database
+from .utils import get_resource_path, tal2mni
 
 
 __all__ = ['convert_sleuth', 'convert_sleuth_to_database',
